Functions_done

+++ dec2/dec2/Entities/BirthProbability.cs
@@ -9,7 +9,7 @@ namespace dec2.Entities
    public class BirthProbability
     {
         public int Age { get; set; }
-        public double NbrOfChildren { get; set; }
+        public int NbrOfChildren { get; set; }
         public double Probability { get; set; }
     }
 }
